Update README.md to mention that package is deprecated (#8)
This metadata package has been deprecated and not maintained since
`aiida-core==1.0`. The README.md doesn't mention this and the package on
PyPI still looks like it is active. It is best to make it more explicit
that this package is deprecated and make one final release to PyPI and
then archive this repository.

The `setup.json` is updated adding `Development Status :: 7 - Inactive`
as a trove classifier and the version limit on `aiida-core` is removed
such that when this package installed through `pip`, at least the latest
version of `aiida-core` gets installed.